added static_cast

+++ src/operator/subgraph/subgraph_property.h
@@ -343,7 +343,7 @@ class SubgraphProperty {
                                     std::vector<std::vector<nnvm::NodeEntry*>>* output_map) const {
     for (size_t i = 0; i < output_map->size(); ++i) {
       for (auto e : output_map[i]) 
-        *e = nnvm::NodeEntry{subgraph_node, i, 0};
+        *e = nnvm::NodeEntry{subgraph_node, static_cast<uint32_t>(i), 0};
     }
   }
   /*!
